Updated Product Controller
