+++ Controllers/ProductController.cs
@@ -1,4 +1,4 @@
-/*using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -12,51 +12,24 @@ using RoastHiveMvc.Models;
 using RoastHiveMvc.Data;
 
 namespace RoastHiveMvc.Controllers;
-
-namespace RoastHiveMvc.Controllers
-{
+    
     [Route("api/[controller]")]
     [ApiController]
     public class ProductController : Controller
     {
-        private readonly ProductsDbContext _context;
+        private readonly RoastHiveDbContext _context;
 
-        public ProductController(ProductsDbContext context)
+        public ProductController(RoastHiveDbContext context)
         {
             _context = context;
         }
 
-        // GET: Product
-        public async Task<IActionResult> Index()
+        // GET: Product Details
+        public async Task<IActionResult> Details(int? id)
         {
-              return _context.Product != null ? 
-                          View(await _context.Product.ToListAsync()) :
-                          Problem("Entity set 'ApplicationDbContext.Product'  is null.");
-        }
-
-        // GET: Product/ShowSearchForm
-
-        public async Task<IActionResult> ShowSearchForm()
-        {
-            return _context.Product != null ?
-                        View() :
-                        Problem("Entity set 'ProductsDbContext.Product'  is null.");
-        }
-        // POST: Product/ShowSearchResults
-        
-        public async Task<IActionResult> ShowSearchResults(String SearchPhrase)
-        {
-            return View("Index", await _context.Product.Where(j => j.Name.Contains
-            (SearchPhrase)).ToListAsync());
-        }
-        
-
-        // GET: Product/Details/5
-        public async Task<IActionResult> Description (int? id)
-        {
-            if (id == null || _db.Product == null)
+            if (id == null || _context.Product == null)
             {
-                return NotFound();
+                return NotFound(); 
             }
 
             var product = await _context.Product
@@ -68,135 +41,5 @@ namespace RoastHiveMvc.Controllers
 
             return View(product);
         }
-
-        // GET: Product/Create
-        [Authorize]
-        public IActionResult Create()
-        {
-            return View();
-        }
-        [Authorize]
-
-        // POST: Product/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ProdID, CatId, Name, Description, Size, UnitPrice, Origin")] Product product)
-        {
-            if (ModelState.IsValid)
-            {
-                _context.Add(product);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(product);
-        }
         
-        [Authorize]
-
-        // GET: Product/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null || _context.Product == null)
-            {
-                return NotFound();
-            }
-
-            var product = await _context.Product.FindAsync(id);
-            if (product == null)
-            {
-                return NotFound();
-            }
-            return View(product);
-        }
-
-        [Authorize]
-        // POST: Product/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ProdID, CatId, Name, Description, Size, UnitPrice, Origin")] Product product)
-        {
-            if (id != product.ProdID)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(product);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ProductExists(product.ProdID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(product);
-        }
-
-        private bool ProductExists(int? id)
-        {
-            throw new NotImplementedException();
-        }
-
-        [Authorize]
-
-        // GET: Product/Delete/5
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null || _context.Product == null)
-            {
-                return NotFound();
-            }
-
-            var product = await _context.Product
-                .FirstOrDefaultAsync(m => m.ProdID == id);
-            if (product == null)
-            {
-                return NotFound();
-            }
-
-            return View(product);
-        }
-        [Authorize]
-
-        // POST: Product/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            if (_context.Product == null)
-            {
-                return Problem("Entity set 'ProductsDbContext.Product'  is null.");
-            }
-            var product = await _context.Product.FindAsync(id);
-            if (product != null)
-            {
-                _context.Product.Remove(product);
-            }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
-        }
-
-        private bool ProductExists(int id)
-        {
-          return (_context.Product?.Any(e => e.ProdID == id)).GetValueOrDefault();
-        }
     }
-}
-
-*/
